chore: rename references to `block` which no longer makes sense
