+++ acir/src/circuit/opcodes/memory_operation.rs
